requirements.txt
deepagents==0.4.1
streamlit
requests
PyYAML

langgraph
langchain-core
langchain-openai
langchain-community
langchain-chroma
langchain-text-splitters
chromadb
pypdf

dashscope
